make AgeCheck container
related #5
