+++ app/components/AgeInfo.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const AgeInfo = () => {
+  return (
+    <div className='age-info'>
+      <h3>What can I do if I am over 18?</h3>
+      <p>
+        If you're 18 or older, contact the website(s) where you've seen the image or video, or report it to the police.
+      </p>
+    </div>
+  )
+}
+
+export default AgeInfo
