+++ app/index.js
@@ -12,7 +12,7 @@ import forms from './reducers/forms'
 import muiTheme from './assets/theme'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 // Components
-import AgeCheck from './components/AgeCheckPage'
+import AgeCheck from './containers/AgeCheck'
 import Home from './containers/Home'
 import Forms from './containers/Forms'
 import OverAge from './components/OverAge'
